requirements.txt
python-magic
pygobject
pycairo
